+++ apps/api/v2/src/modules/billing/billing.module.ts
@@ -1,7 +1,7 @@
-import { BillingConfigService } from "@/modules/billing/billing.config.service";
 import { BillingRepository } from "@/modules/billing/billing.repository";
-import { BillingService } from "@/modules/billing/billing.service";
 import { BillingController } from "@/modules/billing/controllers/billing.controller";
+import { BillingConfigService } from "@/modules/billing/services/billing.config.service";
+import { BillingService } from "@/modules/billing/services/billing.service";
 import { MembershipsModule } from "@/modules/memberships/memberships.module";
 import { OrganizationsModule } from "@/modules/organizations/organizations.module";
 import { PrismaModule } from "@/modules/prisma/prisma.module";
